gate(m-infra): verify-live-virtio-e2e.sh — input in, structured console + snapshot out over custom virtio (claim 0680)
Live PASS on macOS 27 VZ: 12 HID messages typed over queue 3 while the
terminal owns focus; the typed command's report (events=6, armed=0) read
back through the queue-1 structured console FILE; one framebuffer snapshot
streamed as 113 checksummed chunks over queue 4 into a byte-exact
3686400-byte raw BGRX frame; five-queue attach; no synthesis/SCK anywhere.
Regressions: input(GATE_VIRTIO=1)/pointer-virtio/cvc-echo/custom-virtio all
PASS unchanged.

Two live findings fixed en route:
- polled q1 sends leaked their descriptor chains (spike's 5 lines survived;
  sustained tee traffic exhausted the 32-ring) — cvlog_puts/tee/snapshot
  now defer free_chain_q;
- RFC-1071 needs a u64 accumulator at whole-frame scale (u32 overflowed as
  a host Swift trap mid-stream); both sides fixed + overflow test vector.

+++ host/vm-runner/Sources/VMRunner/main.swift
@@ -4330,14 +4330,17 @@ final class CustomVirtioSpikeDeviceDelegate: NSObject, VZCustomVirtioDeviceDeleg
     /// RFC 1071 one's-complement Internet checksum over big-endian words —
     /// byte-for-byte the same function as the guest's ipv4.checksum /
     /// virtio_custom.checksum1071, so the host verifies what the guest sent.
+    /// The accumulator is UInt64: a whole-frame checksum over the 3.5 MiB
+    /// scanout overflows UInt32 before the fold (observed live as a Swift
+    /// "arithmetic overflow" trap, claim 0680).
     private func rfc1071(_ b: [UInt8]) -> UInt16 {
-        var sum: UInt32 = 0
+        var sum: UInt64 = 0
         var i = 0
         while i + 1 < b.count {
-            sum += (UInt32(b[i]) << 8) | UInt32(b[i + 1])
+            sum += (UInt64(b[i]) << 8) | UInt64(b[i + 1])
             i += 2
         }
-        if i < b.count { sum += UInt32(b[i]) << 8 }
+        if i < b.count { sum += UInt64(b[i]) << 8 }
         while (sum >> 16) != 0 { sum = (sum & 0xffff) + (sum >> 16) }
         return ~UInt16(sum & 0xffff)
     }
